+++ src/Model/Catalog/Product/Gallery/Processor.php
@@ -105,7 +105,8 @@ class Processor extends ParentProcessor
             'position' => $position,
             'label' => '',
             'disabled' => (int)$exclude,
-            'media_type' => 'image'
+            'media_type' => 'image',
+            'types'      => $mediaAttribute
         ];
 
         $product->setData($attrCode, $mediaGalleryData);