+++ src/Model/Catalog/ProductRepository.php
@@ -17,12 +17,155 @@ namespace DieKavallerie\ProductImageFix\Model\Catalog;
 use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Catalog\Model\ProductRepository as ParentProductRepository;
 
+use Magento\Framework\Api\Data\ImageContentInterface;
+use Magento\Framework\Exception\InputException;
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Exception\StateException;
+use Magento\Framework\Exception\ValidatorException;
+use Magento\Framework\Exception\CouldNotSaveException;
+
 /**
  * Class ProductRepository
  * @package DieKavallerie\ProductImageFix\Model\Catalog
  */
 class ProductRepository extends ParentProductRepository
 {
+    /**
+     * {@inheritdoc}
+     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
+     * @SuppressWarnings(PHPMD.NPathComplexity)
+     */
+    public function save(\Magento\Catalog\Api\Data\ProductInterface $product, $saveOptions = false)
+    {
+        $tierPrices = $product->getData('tier_price');
+
+        try {
+            $existingProduct = $this->get($product->getSku());
+
+            $product->setData(
+                $this->resourceModel->getLinkField(),
+                $existingProduct->getData($this->resourceModel->getLinkField())
+            );
+        } catch (NoSuchEntityException $e) {
+            $existingProduct = null;
+        }
+
+        $productDataArray = $this->extensibleDataObjectConverter
+            ->toNestedArray($product, [], 'Magento\Catalog\Api\Data\ProductInterface');
+        $productDataArray = array_replace($productDataArray, $product->getData());
+        $ignoreLinksFlag  = $product->getData('ignore_links_flag');
+        $productLinks     = null;
+
+        if (!$ignoreLinksFlag && $ignoreLinksFlag !== null) {
+            $productLinks = $product->getProductLinks();
+        }
+
+        $productDataArray['store_id'] = (int)$this->storeManager->getStore()->getId();
+        $product                      = $this->initializeProductData($productDataArray, empty($existingProduct));
+
+        $this->processLinks($product, $productLinks);
+        if (isset($productDataArray['media_gallery']['images'])) {
+            $mediaGalleryEntries = $productDataArray['media_gallery']['images'];
+            unset($productDataArray['media_gallery']);
+
+            $this->processMediaGallery($product, $mediaGalleryEntries);
+        }
+
+        if (!$product->getOptionsReadonly()) {
+            $product->setCanSaveCustomOptions(true);
+        }
+
+        $validationResult = $this->resourceModel->validate($product);
+        if (true !== $validationResult) {
+            throw new \Magento\Framework\Exception\CouldNotSaveException(
+                __('Invalid product data: %1', implode(',', $validationResult))
+            );
+        }
+
+        try {
+            if ($tierPrices !== null) {
+                $product->setData('tier_price', $tierPrices);
+            }
+            unset($this->instances[$product->getSku()]);
+            unset($this->instancesById[$product->getId()]);
+            $this->resourceModel->save($product);
+        } catch (\Magento\Eav\Model\Entity\Attribute\Exception $exception) {
+            throw \Magento\Framework\Exception\InputException::invalidFieldValue(
+                $exception->getAttributeCode(),
+                $product->getData($exception->getAttributeCode()),
+                $exception
+            );
+        } catch (ValidatorException $e) {
+            throw new CouldNotSaveException(__($e->getMessage()));
+        } catch (LocalizedException $e) {
+            throw $e;
+        } catch (\Exception $e) {
+            throw new \Magento\Framework\Exception\CouldNotSaveException(__('Unable to save product'));
+        }
+        unset($this->instances[$product->getSku()]);
+        unset($this->instancesById[$product->getId()]);
+        return $this->get($product->getSku());
+    }
+
+    /**
+     * Process product links, creating new links, updating and deleting existing links
+     *
+     * @param \Magento\Catalog\Api\Data\ProductInterface $product
+     * @param \Magento\Catalog\Api\Data\ProductLinkInterface[] $newLinks
+     * @return $this
+     * @throws NoSuchEntityException
+     */
+    private function processLinks(\Magento\Catalog\Api\Data\ProductInterface $product, $newLinks)
+    {
+        if ($newLinks === null) {
+            // If product links were not specified, don't do anything
+            return $this;
+        }
+
+        // Clear all existing product links and then set the ones we want
+        $linkTypes = $this->linkTypeProvider->getLinkTypes();
+        foreach (array_keys($linkTypes) as $typeName) {
+            $this->linkInitializer->initializeLinks($product, [$typeName => []]);
+        }
+
+        // Set each linktype info
+        if (!empty($newLinks)) {
+            $productLinks = [];
+            foreach ($newLinks as $link) {
+                $productLinks[$link->getLinkType()][] = $link;
+            }
+
+            foreach ($productLinks as $type => $linksByType) {
+                $assignedSkuList = [];
+                /** @var \Magento\Catalog\Api\Data\ProductLinkInterface $link */
+                foreach ($linksByType as $link) {
+                    $assignedSkuList[] = $link->getLinkedProductSku();
+                }
+                $linkedProductIds = $this->resourceModel->getProductsIdsBySkus($assignedSkuList);
+
+                $linksToInitialize = [];
+                foreach ($linksByType as $link) {
+                    $linkDataArray = $this->extensibleDataObjectConverter
+                        ->toNestedArray($link, [], 'Magento\Catalog\Api\Data\ProductLinkInterface');
+                    $linkedSku = $link->getLinkedProductSku();
+                    if (!isset($linkedProductIds[$linkedSku])) {
+                        throw new NoSuchEntityException(
+                            __('Product with SKU "%1" does not exist', $linkedSku)
+                        );
+                    }
+                    $linkDataArray['product_id'] = $linkedProductIds[$linkedSku];
+                    $linksToInitialize[$linkedProductIds[$linkedSku]] = $linkDataArray;
+                }
+
+                $this->linkInitializer->initializeLinks($product, [$type => $linksToInitialize]);
+            }
+        }
+
+        $product->setProductLinks($newLinks);
+        return $this;
+    }
+
     /**
      * @param ProductInterface $product
      * @param array $mediaGalleryEntries
@@ -38,7 +181,9 @@ class ProductRepository extends ParentProductRepository
         if (!empty($existingMediaGallery)) {
             $entriesById = [];
             foreach ($mediaGalleryEntries as $entry) {
-                if (isset($entry['id'])) {
+                if (isset($entry['value_id'])) {
+                    $entriesById[$entry['value_id']] = $entry;
+                } elseif (isset($entry['id'])) {
                     $entry['value_id'] = $entry['id'];
                     $entriesById[$entry['value_id']] = $entry;
                 } else {
@@ -46,8 +191,13 @@ class ProductRepository extends ParentProductRepository
                 }
             }
             foreach ($existingMediaGallery as $key => &$existingEntry) {
+
+                if (!isset($existingEntry['value_id'])) {
+                    continue;
+                }
+
                 if (isset($entriesById[$existingEntry['value_id']])) {
-                    $updatedEntry = $entriesById[$existingEntry['value_id']];
+                    $updatedEntry               = $entriesById[$existingEntry['value_id']];
                     $existingMediaGallery[$key] = array_merge($existingEntry, $updatedEntry);
                 } else {
                     //set the removed flag
@@ -69,14 +219,10 @@ class ProductRepository extends ParentProductRepository
             }
         }
 
-        // The processing of already ( newly ) set images is very buggy. If we had an correctly set media_gallery entry in product it will be removed and reset in save method.
-        // But here it seems as only base 64 encoded images are wanted so we need to re set the ( already ) correct media_gallery array.
-        $mediaGalleryData['images'] = [];
         foreach ($newEntries as $newEntry) {
 
             if (isset($newEntry['file'])) {
                 $this->getMediaGalleryProcessor()->setMediaAttribute($product, $newEntry['types'], $newEntry['file']);
-                $mediaGalleryData['images'][] = $newEntry;
                 continue;
             }
 
@@ -90,11 +236,6 @@ class ProductRepository extends ParentProductRepository
                 ->setType($newEntry['content'][ImageContentInterface::TYPE]);
             $newEntry['content'] = $contentDataObject;
             $this->processNewMediaGalleryEntry($product, $newEntry);
-        }
-
-        // Just set the missing data again
-        if (count($mediaGalleryData['images']) > 0) {
-            $product->setData($this->getMediaGalleryProcessor()->getAttribute()->getAttributeCode(), $mediaGalleryData);
         }
 
         return $this;